Add no-backup option to setup-environment script

+++ setup-environment.sh
@@ -14,22 +14,24 @@ function updateFile() {
 		echo "ERRIR: Cannot update ${destination}: ${source} doesnt exist." 1>&2
 	fi 
 
-	if [ -f "${destination}" ]; then
-		echo "update  ${destination} with ${source}"
+	if [ -f "${destination}" ] && [ "${NO_BACKUP}" == "1" ]; then
 		cp -L "${destination}" "${destination}$(date).bkp"
 		echo "Backed up ${destination} at ${destination}$(date).bkp"
 	fi
+	echo "update  ${destination} with ${source}"
 	ln -sf "${source}" "${destination}"
 }
 
-VAR_ARGS=$(getopt -o o: -l output-dir: -- "$@")
+VAR_ARGS=$(getopt -o o:n -l output-dir:no-backup -- "$@")
 echo "${VAR_ARGS}"
 eval set -- "${VAR_ARGS}"
 
 OUTPUT_DIR=$(realpath "${HOME}")
+NO_BACKUP="1"
 while true; do
 	case "$1" in
 		-o | --output-dir) OUTPUT_DIR=$(realpath "$2"); shift 2;;
+		-n | --no-baskup) NO_BACKUP="0"; shift 1;;
 		*) break;;
 	esac
 done
